refactor(t-61): split ci/mod.rs into per-action submodules (byte-identical)
src/mcp/handlers/ci/mod.rs was long-term oversized (1420 LOC, KNOWN_OVERSIZED
ceiling 1435). Split by action group into 5 sibling submodules, all under MAX_LOC
(750); mod.rs drops to 116 LOC. Pure mechanical move — handler bodies are
byte-identical, only `use`/`pub` adjusted and the cross-module references re-pathed
through `super::`.

- ci/checkout.rs (319) — checkout_source, handle_checkout_repo[_inner], log_checkout_outcome
- ci/watch.rs    (469) — handle_watch_ci / handle_unwatch_ci / handle_status_ci, build_default_provider
- ci/merge.rs    (277) — MergeVerdict, classify_merge_summary, verify_merge_landed, base_drift_refusal, handle_merge_repo
- ci/cleanup.rs  (154) — handle_cleanup_init_commits, handle_cleanup_merged_branches
- ci/release.rs  (119) — validate_release_path, handle_release_repo
- mod.rs keeps `resolve_repo_or_error` (shared by watch + merge, called via super::)
  + the sub-module decls + re-exports.

Interface preserved by re-exports — every `ci::handle_*` path (dispatch.rs) and every
`super::*` path (the child `tests` module, which does `use super::*`) resolves
unchanged. ZERO caller/test edits.

Invariant sync (CI hard-gates these):
- file_size_invariant.rs: removed the ci/mod.rs KNOWN_OVERSIZED entry (now <= MAX_LOC).
- invariant_inputs.rs: removed ci/mod.rs from GRANDFATHERED_OVERSIZED_HANDLERS (kept in
  lock-step with KNOWN_OVERSIZED per the cross-check test).
- daemon_git_helper_invariant.rs: moved the deliberate-raw-git seal from ci/mod.rs to
  ci/release.rs (the worktree-remove travels with handle_release_repo, marker intact).
- merge_freshness.rs tests: swapped the sample grandfathered-file fixture ci/mod.rs →
  dispatch_hook/mod.rs (the remaining grandfathered entry).

Verified: ci handler suite 76/76, p0b+dispatch 36/36, merge_freshness 8/8, file_size 4/4,
git_helper 2/2 — all green; clippy --tests --features tray -D warnings + fmt clean;
cargo build clean.


Agend-Agent: fixup-dev-3
Agend-Task: t-20260615054511406547-61
Agend-Branch: refactor/ci-mod-split
Agend-Issued-At: 2026-06-16T04:59:46.052273+00:00
